tanx/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
decorator==5.1.1
Django==4.2.11
django-cors-headers==4.3.1
django-crontab==0.7.1
django-redis==5.4.0
django-rq==2.10.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
multidict==6.0.5
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.20.0
pygments==2.17.2
PyJWT==1.7.1
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
regex==2023.12.25
requests==2.31.0
rq==1.16.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing-extensions==4.10.0
tzlocal==5.2
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
